Change handling of reserved fields in `PciInfo.try_into_c()`
Set the reserved fields to `0` instead of `u32::MAX` and stopped checking that they match in the `pci_info_from_to_c` test.

Also set the crate `recursion_limit` back to `1024` for the time being (beta should have that as the default but it doesn't seem to have been changed yet?).